Updated filter spec to handle multiple outputs for a single event
This might happen when the Logstash `split` function is used to create
multiple events from a single event based on a particular field.

Rather than just taking `result[0]` in the filter spec, it now loops
through all the results returned and compares them to the expected
output as defined in the JSON test file. When Logstash splits a single
event into multiple events, most of the event fields are copied from the
original to all the newly generated events. As such, the JSON test file
should identify the fields that will have multiple results via an array
of the results.

For example:

    "cases": [{
      "in": "....",
      "out": {
        "type": "foo", # same for all events
        "message": ["bar", "fish"] # two events should exist
      }
    }]